@@ -25,8 +25,15 @@ public final class DMBeanUtils {
 				String type = opMetadata.getType();
 				int impact = opMetadata.getImpact();
 
-				MBeanOperationInfo opInfo = new MBeanOperationInfo(name, desc, /* signature */ null, type, impact, /* descriptor */ null);
-				ops[i++] = opInfo;
+				try {
+					MBeanOperationInfo opInfo = new MBeanOperationInfo(name, desc, /* signature */ null, type, impact, /* descriptor */ null);
+					ops[i++] = opInfo;
+				} catch (IllegalArgumentException e) {
+					logger.log(Level.WARNING, "MBeanOperationInfo.UNKNOWN is used as the value for the impact field");
+					impact = MBeanOperationInfo.UNKNOWN;
+					MBeanOperationInfo opInfo = new MBeanOperationInfo(name, desc, /* signature */ null, type, impact, /* descriptor */ null);
+					ops[i++] = opInfo;
+				}
 			}
 		}
 		
